requirements.txt
lgpio==0.2.2.0
rpi-lgpio==0.6